Add Proc as a type for ActiveJob retry_on attempts
ActiveJob retry_on allowed the wait option to be set as a Proc
that calculates the wait but did not give the same option for
attempts. attempts could be set to only a number of times to
attempt or unlimited. However, jobs sometimes need different
logic than number of executions to determine if they should
try again.

This commit adds support for setting attempts to a Proc that
returns either true or false. On true the retry will occur.

As a an example:

    retry_on ProcRetryError, attempts: ->(executions) { executions < 2 }